+++ packages/ruleset-migrator/src/__tests__/__fixtures__/formats-variant-2/output.mjs
@@ -1,11 +1,11 @@
-import { asyncapi2, jsonSchemaLoose, oas2, oas3, oas3_0, oas3_1 } from '@stoplight/spectral-formats';
+import { asyncapi2, asyncapi3, jsonSchemaLoose, oas2, oas3, oas3_0, oas3_1 } from '@stoplight/spectral-formats';
 import { truthy } from '@stoplight/spectral-functions';
 export default {
   formats: [oas2, oas3_1, oas3_0, jsonSchemaLoose],
   rules: {
     test: {
       given: '$',
-      formats: [asyncapi2, oas3, oas3_0, oas3_1],
+      formats: [asyncapi2, asyncapi3, oas3, oas3_0, oas3_1],
       then: {
         function: truthy,
       },