+++ packages/rulesets/src/asyncapi/functions/utils/specs.ts
@@ -1,4 +1,8 @@
-import specs from '@asyncapi/specs';
+/* eslint-disable @typescript-eslint/no-unsafe-argument */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+import allSchemas from '@asyncapi/specs';
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+const specs = allSchemas.schemas;
 
 export type AsyncAPISpecVersion = keyof typeof specs;